+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [1.1.2] - November 13, 2021
+
+* Fixed bug in RenderFloatColumn `visitChildrenOfAnyType`.
+* Added example Mac app, and updated examples.
+
 ## [1.1.1] - September 9, 2021
 
 * Updated to use `flutter_lints_plus` package and fix Flutter 2.5 lint warnings.